#3 Weight definition

+++ IRProject/src/main/java/hashing/SimHash.java
@@ -1,6 +1,6 @@
 package hashing;
 
 public interface SimHash {
-	public int transform(String inputString);
+	public long transform(String inputString);
 
 }
